api integrate 17-04-2024

+++ lib/app/modules/my_card/controllers/my_card_controller.dart
@@ -4,6 +4,7 @@ import 'package:HibaPay/app/data/apis/api_models/delete_virtual_card_model.dart'
 import 'package:HibaPay/app/data/apis/api_models/get_card_holder_model.dart';
 import 'package:HibaPay/app/data/apis/api_models/list_virtual_cards_model.dart';
 import 'package:HibaPay/app/routes/app_pages.dart';
+import 'package:HibaPay/common/common_widgets.dart';
 import 'package:get/get.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -48,22 +49,29 @@ class MyCardController extends GetxController {
   }
 
   clickOnDeleteButton({required int index}) async {
-    if (listVirtualCardsResult[index].vcardCardId != null &&
-        listVirtualCardsResult[index].vcardCardId!.isNotEmpty &&
-        authTokenHiba.value.isNotEmpty) {
-      inAsyncCall.value = true;
-      bodyParams = {
-        ApiKeyConstants.authTokenHiba: authTokenHiba.value,
-        ApiKeyConstants.virtualCardId:
-            listVirtualCardsResult[index].vcardCardId ?? ''
-      };
-      DeleteVirtualCardModel? deleteVirtualCardModel =
-          await ApiMethods.deleteVirtualCard(bodyParams: bodyParams);
-      if (deleteVirtualCardModel != null &&
-          deleteVirtualCardModel.result != null) {
-        await onInitWork();
-      }
-    }
+    CommonWidgets.showAlertDialog(
+      title: 'Delete Info',
+      content: 'Delete your virtual card',
+      onPressedYes: () async {
+        Get.back();
+        if (listVirtualCardsResult[index].vcardCardId != null &&
+            listVirtualCardsResult[index].vcardCardId!.isNotEmpty &&
+            authTokenHiba.value.isNotEmpty) {
+          inAsyncCall.value = true;
+          bodyParams = {
+            ApiKeyConstants.authTokenHiba: authTokenHiba.value,
+            ApiKeyConstants.virtualCardId:
+                listVirtualCardsResult[index].vcardCardId ?? ''
+          };
+          DeleteVirtualCardModel? deleteVirtualCardModel =
+              await ApiMethods.deleteVirtualCard(bodyParams: bodyParams);
+          if (deleteVirtualCardModel != null &&
+              deleteVirtualCardModel.result != null) {
+            await onInitWork();
+          }
+        }
+      },
+    );
   }
 
   getCardHolderApi() async {
